rename package.js to create-homebrew-pr.js
