+++ .github/workflows/release.yml
@@ -37,3 +37,4 @@ jobs:
           with: |
             version: pnpm changeset-manifests
             title: Version Packages - ${{ github.ref_name }}
+            publish: pnpm release latest
