+++ bin/get-graphql-schemas.js
@@ -3,6 +3,8 @@ import {Octokit} from '@octokit/rest'
 import * as fs from 'fs'
 import * as path from 'path'
 import {spawn} from 'child_process'
+import {withOctokit} from './github-utils.js'
+import {runCommand} from './run-command.js'
 
 const BRANCH = 'main'
 
@@ -71,46 +73,6 @@ const schemas = [
   },
 ]
 
-function runCommand(command, args) {
-  return new Promise((resolve, reject) => {
-    const child = spawn(command, args, {stdio: ['inherit', 'pipe', 'pipe']})
-
-    let output = ''
-    let errorOutput = ''
-
-    child.stdout.on('data', (data) => {
-      console.log(data.toString())
-      output += data.toString()
-    })
-
-    child.stderr.on('data', (data) => {
-      console.log(data.toString())
-      errorOutput += data.toString()
-    })
-
-    child.on('close', (code) => {
-      if (code !== 0) {
-        reject(new Error(`Command failed with exit code ${code}\n${errorOutput}`))
-      } else {
-        resolve(output)
-      }
-    })
-  })
-}
-
-/**
- * @param {string} output
- * @returns {string}
- */
-function extractPassword(output) {
-  const passwordRegex = /Password: (\w+)/
-  const match = output.match(passwordRegex)
-  if (match && match[1]) {
-    return match[1]
-  }
-  throw new Error('Password not found in output')
-}
-
 /**
  * @param {Schema} schema
  * @param {import('@octokit/rest').Octokit} octokit
@@ -163,55 +125,6 @@ async function fetchFileForSchema(schema, octokit) {
     console.error(`Error fetching file: ${error.message}`)
     return false
   }
-}
-
-/**
- * @returns {Promise<string>}
- */
-async function getGithubPasswordFromDev() {
-  try {
-    // Uses token from `dev`
-    const output = await runCommand('/opt/dev/bin/dev', ['github', 'print-auth'])
-    const password = extractPassword(output)
-    return password
-  } catch (error) {
-    console.warn(`Soft-error fetching password from dev: ${error.message}`)
-    process.exit(0)
-  }
-}
-
-/**
- * @param {string} owner
- * @param {function(import('@octokit/rest').Octokit): Promise<boolean>} func
- * @returns {Promise<boolean>}
- */
-async function withOctokit(owner, func) {
-  let password = undefined
-
-  const tokenEnvSources = [
-    `GITHUB_TOKEN_${owner.toUpperCase()}`,
-    `GH_TOKEN_${owner.toUpperCase()}`,
-    'GITHUB_TOKEN',
-    'GH_TOKEN',
-  ]
-  let tokenFromEnv = undefined
-  for (const source of tokenEnvSources) {
-    if (process.env[source]) {
-      tokenFromEnv = process.env[source]
-      console.log(`Using token from ${source}: ${tokenFromEnv}`)
-      break
-    }
-  }
-  if (!tokenFromEnv) {
-    password = await getGithubPasswordFromDev()
-    console.log(`Using password from dev: ${password}`)
-  }
-  const authToken = password || tokenFromEnv
-
-  const octokit = new Octokit({
-    auth: authToken,
-  })
-  return func(octokit)
 }
 
 /**
