+++ package.json
@@ -8,7 +8,6 @@
     "build:affected": "nx affected --target=build",
     "build": "nx run-many --target=build --all --skip-nx-cache",
     "bundle-for-release": "nx run-many --target=bundle --all --skip-nx-cache",
-    "changeset-build-no-docs": "changeset version && pnpm install --no-frozen-lockfile && pnpm build && bin/update-cli-kit-version.js",
     "changeset-manifests": "changeset version && pnpm install --no-frozen-lockfile && pnpm refresh-manifests && pnpm refresh-readme && pnpm refresh-code-documentation && bin/update-cli-kit-version.js",
     "clean": "nx run-many --target=clean --all --skip-nx-cache && nx reset",
     "create-app": "nx build create-app && node packages/create-app/bin/dev.js --package-manager npm",
@@ -21,12 +20,13 @@
     "lint:fix:affected": "nx affected --target=lint:fix",
     "lint:fix": "nx run-many --target=lint:fix --all --skip-nx-cache",
     "lint": "nx run-many --target=lint --all --skip-nx-cache",
-    "package": "bin/package.js",
+    "create-homebrew-pr": "bin/create-homebrew-pr.js",
     "refresh-code-documentation": "nx run-many --target=refresh-code-documentation --all --skip-nx-cache",
     "refresh-manifests": "nx run-many --target=refresh-manifests --all --skip-nx-cache && bin/prettify-manifests.js && pnpm refresh-readme",
     "refresh-readme": "nx run-many --target=refresh-readme --all --skip-nx-cache",
     "refresh-templates": "nx run-many --target=refresh-templates --all --skip-nx-cache",
     "release": "./bin/release",
+    "post-release": "./bin/post-release",
     "shopify:run": "node packages/cli/bin/dev.js",
     "shopify": "nx build cli && node packages/cli/bin/dev.js",
     "test:affected": "nx affected --target=test",
@@ -45,7 +45,6 @@
     "@graphql-codegen/near-operation-file-preset": "3.0.0",
     "@nx/eslint-plugin": "19.8.14",
     "@nx/workspace": "19.8.14",
-    "@octokit/core": "^4.2.4",
     "@octokit/rest": "20.1.1",
     "@shopify/eslint-plugin-cli": "file:packages/eslint-plugin-cli",
     "@shopify/generate-docs": "0.15.6",
