+++ shipit.nightly.yml
@@ -25,4 +25,4 @@ deploy:
     - bash -i -c "npm_config_loglevel=verbose NODE_ENV=production pnpm bundle-for-release"
     - bash -i -c "npm_config_loglevel=verbose node bin/create-cli-duplicate-package.js"
     - bash -i -c "npm_config_loglevel=verbose pnpm changeset publish --tag nightly"
-    - bash -i -c "./bin/package.js"
+    - bash -i -c "./bin/create-homebrew-pr.js"
