+++ shipit.production.yml
@@ -15,5 +15,5 @@ deploy:
     - bash -i -c "npm_config_loglevel=verbose NODE_ENV=production pnpm bundle-for-release"
     - bash -i -c "npm_config_loglevel=verbose node bin/create-cli-duplicate-package.js"
     - bash -i -c "npm_config_loglevel=verbose pnpm changeset publish"
-    - bash -i -c "./bin/package.js"
+    - bash -i -c "./bin/create-homebrew-pr.js"
     - bash -i -c "./bin/create-doc-pr.js"
